move general musicview prefs to own generic page, with print prefs
closes #1198
