+++ frescobaldi_app/pagedview.py
@@ -265,7 +265,7 @@ def getPopplerBackends():
         renderBackend = popplerqt5.Poppler.Document.ArthurBackend
     else:
         renderBackend = popplerqt5.Poppler.Document.SplashBackend
-    if QSettings().value("musicview/arthurbackend_print", False, bool):
+    if QSettings().value("musicview/arthurbackend_print", True, bool):
         printRenderBackend = popplerqt5.Poppler.Document.ArthurBackend
     else:
         printRenderBackend = popplerqt5.Poppler.Document.SplashBackend
